apps/api/pyproject.toml
[project]
name = "scale-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.29.0",
  "alembic>=1.13.3",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.12",
  "slowapi>=0.1.9",
  "httpx>=0.27.0",
  "greenlet>=3.5.0",
  "bcrypt<4.1",
  "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "testcontainers[postgres]>=4.8.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
  "factory-boy>=3.3.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
addopts = "--cov=src/scale_api --cov-report=term-missing --cov-fail-under=80"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "psycopg[binary]>=3.3.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/scale_api"]
